rename orderServies and fix func reqs to class
