+++ src/utils/api/orderService/orderService.ts
@@ -1,47 +1,48 @@
 export interface OrderedMeal {
-    /**
-     * Meal's id
-     */
+    /** Meal's id */
     id: string;
-    /**
-     * Meal's name
-     */
+
+    /** Meal's name */
     name: string;
-    /**
-     * Meal's type
-     */
+
+    /** Meal's type */
     price: number;
-    /**
-     * Number of units of the item ordered
-     */
+
+    /** Number of units of the item ordered */
     quantity: number;
 }
 
-/**
- * Describes the state of an order.
- * @interface
- * @property {string} clientId - Unique identifier for the client.
- * @property {string} id - Unique identifier for the order.
- * @property {OrderedMeal[]} orderedMeal - Array with details of items in the order.
- * @property {number} totalAmount - Total amount of the order.
- * @property {'waiting' | 'confirmed' | 'notConfirmed'} confirmationStatus - Confirmation status of the order.
- * @property {'waiting' | 'reviewed' | 'skipped'} reviewStatus - Review status of the order.
- * @property {number} cancellationTime - Time of order cancellation in seconds.
- * @property {'none' | 'requested' | 'confirmed'} cancellationStatus - Cancellation status of the order.
- * @property {boolean} isCancellationRequested - Flag indicating whether a cancellation was requested for the order.
- * @property {number} preparationTime - Time required to prepare the order, in minutes.
- */
-
 export interface OrderState {
+    /** Unique identifier for the client */
     userId: string;
+
+    /** Unique identifier for the order */
     id: string;
+
+    /** Total amount of the order */
     totalAmount: number;
+
+    /** Confirmation status of the order */
     confirmationStatus: 'waiting' | 'confirmed' | 'notConfirmed';
+
+    /** Time required to prepare the order, in minutes */
     preparationTime: number;
+
+    /** Payment status of the order */
     paymentStatus: string;
-    reviewStatus: string;
+
+    /** Review status of the order */
+    reviewStatus: 'waiting' | 'reviewed' | 'skipped';
+
+    /** Time of order cancellation in seconds */
     cancellationTime: number;
-    cancellationStatus: string;
+
+    /** Cancellation status of the order */
+    cancellationStatus: 'none' | 'requested' | 'confirmed';
+
+    /** Flag indicating whether a cancellation was requested for the order */
     isCancellationRequested: boolean;
+
+    /** Array with details of items in the order */
     orderedMeal: OrderedMeal[];
 }
